requirements.txt
requests
python-decouple
pp-ez
# PIL
# urllib3
# httplib2
# httpx